+++ wrpc.core/Models/ModelsJsonContext.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Text.Json.Serialization;
 using WasabiRpc.Models.App;
