+++ wrpc.core/ViewModels/BatchMode/BatchManagerViewModel.cs
@@ -4,7 +4,6 @@ using System.Threading.Tasks;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using WasabiRpc.Models.BatchMode;
-using WasabiRpc.Models.Results;
 using WasabiRpc.Models.Services;
 
 namespace WasabiRpc.ViewModels.BatchMode;
