+++ wrpc.core/ViewModels/Methods/BuildViewModel.cs
@@ -12,6 +12,7 @@ using WasabiRpc.Models.Params.Build;
 using WasabiRpc.Models.Params.Send;
 using WasabiRpc.Models.Results;
 using WasabiRpc.Models.Services;
+using WasabiRpc.ViewModels.Methods.Adapters;
 
 namespace WasabiRpc.ViewModels.Methods;
 
@@ -50,7 +51,7 @@ public partial class BuildViewModel : RoutableMethodViewModel
     private decimal? _feeRate;
 
     [ObservableProperty]
-    private ObservableCollection<CoinViewModel> _coins;
+    private ObservableCollection<CoinAdapterViewModel> _coins;
 
     public BuildViewModel(IRpcServiceViewModel rpcService, INavigationService navigationService, IBatchManager batchManager, string walletName)
         : base(rpcService, navigationService, batchManager)
@@ -63,7 +64,7 @@ public partial class BuildViewModel : RoutableMethodViewModel
         SubtractFee = false;
         FeeTarget = 2;
         FeeTarget = null;
-        Coins = new ObservableCollection<CoinViewModel>();
+        Coins = new ObservableCollection<CoinAdapterViewModel>();
     }
 
     private bool CanBuild()
@@ -172,9 +173,9 @@ public partial class BuildViewModel : RoutableMethodViewModel
         {
             var coins = rpcListUnspentCoinsResult
                 .Result
-                .Select(x => new CoinViewModel(RpcService, NavigationService, BatchManager, WalletName, x.ToViewModel(RpcService, NavigationService)));
+                .Select(x => new CoinAdapterViewModel(RpcService, NavigationService, BatchManager, WalletName, x.ToViewModel(RpcService, NavigationService)));
 
-            Coins = new ObservableCollection<CoinViewModel>(coins);
+            Coins = new ObservableCollection<CoinAdapterViewModel>(coins);
         }
         else if (result is RpcErrorResult { Error: not null } rpcErrorResult)
         {
